requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-js-asset==2.2.0
djangorestframework==3.15.2
mysqlclient==2.2.6
numpy==2.1.2
opencv-python==4.10.0.84
Pillow==11.0.0
pyzbar==0.1.9
sqlparse==0.5.2
tzdata==2024.2
